fix incorrect file size counting
The code was not printing the correct reducted total token for input
file with TBB parallelization. It might be related to the delcaration
scope of the lambda function. So, we remove and inline to appropreiate
location.

Also, by seperating OMP_TBB implemenation of readText out, we make the
code cleaner.

Signed-off-by: Andy Chiu <andybnac@gmail.com>